Refactor pairing
